sudo animation added
sudo animation added. still need to learn how to actually animate objects in textual. computation is pending. I ended up pushing forward on the animation as the compute was easy to transfer over, but i still need to add it back in again and this time link it to the corresponding buttons -> text fields